+++ model/model_chain_publish_queue_item.go
@@ -14,7 +14,7 @@ type ChainPublishQueueItem struct {
 	UpdateAt string `json:"UpdateAt,omitempty"`
 	// also stored in the key for quick indexing
 	GroupId    string `json:"groupId,omitempty"`
-	RetryCount int32  `json:"retryCount,omitempty"`
+	RetryCount int64  `json:"retryCount,omitempty"`
 	// in value only
 	State       string `json:"state,omitempty"`
 	StorageType string `json:"storageType,omitempty"`
