+++ model/model_timestamppb_timestamp.go
@@ -12,7 +12,7 @@ package model
 // TimestamppbTimestamp struct for TimestamppbTimestamp
 type TimestamppbTimestamp struct {
 	// Non-negative fractions of a second at nanosecond resolution. Negative second values with fractions must still have non-negative nanos values that count forward in time. Must be from 0 to 999,999,999 inclusive.
-	Nanos int32 `json:"nanos,omitempty"`
+	Nanos int64 `json:"nanos,omitempty"`
 	// Represents seconds of UTC time since Unix epoch 1970-01-01T00:00:00Z. Must be from 0001-01-01T00:00:00Z to 9999-12-31T23:59:59Z inclusive.
-	Seconds int32 `json:"seconds,omitempty"`
+	Seconds int64 `json:"seconds,omitempty"`
 }
